# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "## TMF API Reference: TMF633 - Service Catalog Management\n\n### Release\
    \ : 18.5 - December 2018\n\nService Catalog API is one of Catalog Management API\
    \ Family. Service Catalog API goal is to provide a catalog of services. \n\n###\
    \ Resource \n- serviceCatalog\n\n### Operations\nService Catalog API performs\
    \ the following operations on the resource :\n- Retrieve an entity or a collection\
    \ of entities depending on filter criteria\n- Partial update of an entity (including\
    \ updating rules)\n- Create an entity (including default values and creation rules)\n\
    - Delete an entity (for administration purposes)\n- Manage notification of events"
  version: "4.0.0"
  title: "Service Catalog Management"
host: "serverRoot"
basePath: "/tmf-api/serviceCatalogManagement/v4/"
...
schemes:
- "https"
consumes:
- "application/json;charset=utf-8"
produces:
- "application/json;charset=utf-8"
paths:
...
  /serviceCategory:
    get:
      tags:
      - "serviceCategory"
      summary: "List or find ServiceCategory objects"
      description: "This operation list or find ServiceCategory entities"
      operationId: "listServiceCategory"
      parameters:
      - name: "fields"
        in: "query"
        description: "Comma-separated properties to be provided in response"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "Requested index for start of resources to be provided in response"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "Requested number of resources to be provided in response"
        required: false
        type: "integer"
      responses:
        200:
          description: "Success"
          headers:
            X-Total-Count:
              type: "integer"
              description: "Total number of items matching criteria"
            X-Result-Count:
              type: "integer"
              description: "Actual number of items returned in the response body"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ServiceCategory"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceCategory"
    post:
      tags:
      - "serviceCategory"
      summary: "Creates a ServiceCategory"
      description: "This operation creates a ServiceCategory entity."
      operationId: "createServiceCategory"
      parameters:
      - in: "body"
        name: "serviceCategory"
        description: "The ServiceCategory to be created"
        required: true
        schema:
          $ref: "#/definitions/ServiceCategory_Create"
      responses:
        201:
          description: "Created"
          schema:
            $ref: "#/definitions/ServiceCategory"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceCategory"
  /serviceCategory/{id}:
    get:
      tags:
      - "serviceCategory"
      summary: "Retrieves a ServiceCategory by ID"
      description: "This operation retrieves a ServiceCategory entity. Attribute selection\
        \ is enabled for all first level attributes."
      operationId: "retrieveServiceCategory"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the ServiceCategory"
        required: true
        type: "string"
      - name: "fields"
        in: "query"
        description: "Comma-separated properties to provide in response"
        required: false
        type: "string"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/ServiceCategory"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceCategory"
    delete:
      tags:
      - "serviceCategory"
      summary: "Deletes a ServiceCategory"
      description: "This operation deletes a ServiceCategory entity."
      operationId: "deleteServiceCategory"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the ServiceCategory"
        required: true
        type: "string"
      responses:
        204:
          description: "Deleted"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceCategory"
    patch:
      tags:
      - "serviceCategory"
      summary: "Updates partially a ServiceCategory"
      description: "This operation updates partially a ServiceCategory entity."
      operationId: "patchServiceCategory"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the ServiceCategory"
        required: true
        type: "string"
      - in: "body"
        name: "serviceCategory"
        description: "The ServiceCategory to be updated"
        required: true
        schema:
          $ref: "#/definitions/ServiceCategory_Update"
      responses:
        200:
          description: "Updated"
          schema:
            $ref: "#/definitions/ServiceCategory"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceCategory"
...
definitions:
...
  ServiceCandidateRef:
    type: "object"
    required:
    - "href"
    - "id"
    properties:
      id:
        type: "string"
        description: "Unique identifier of the service candidate"
      href:
        type: "string"
        description: "Unique reference of the service candidate"
      name:
        type: "string"
        description: "Name of the service candidate"
      version:
        type: "string"
        description: "Version of the service candidate"
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class"
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name"
      '@referredType':
        type: "string"
        description: "The actual type of the target instance when needed for disambiguation."
    description: "ServiceCandidate reference. ServiceCandidate is an entity that makes\
      \ a ServiceSpecification available to a catalog."
    example:
      '@referredType': "@referredType"
      '@baseType': "@baseType"
      '@type': "@type"
      name: "name"
      id: "id"
      href: "href"
      '@schemaLocation': "http://example.com/aeiou"
      version: "version"
...
  ServiceCategory:
    type: "object"
    properties:
      id:
        type: "string"
        description: "Unique identifier of the category"
      href:
        type: "string"
        description: "Reference of the category"
      description:
        type: "string"
        description: "Description of the category"
      isRoot:
        type: "boolean"
        description: "If true, this Boolean indicates that the category is a root\
          \ of categories"
      lastUpdate:
        type: "string"
        format: "date-time"
        description: "Date and time of the last update"
      lifecycleStatus:
        type: "string"
        description: "Used to indicate the current lifecycle status"
      name:
        type: "string"
        description: "Name of the category"
      parentId:
        type: "string"
        description: "Unique identifier of the parent category"
      version:
        type: "string"
        description: "ServiceCategory version"
      category:
        type: "array"
        description: "List of child categories in the tree for in this category"
        items:
          $ref: "#/definitions/ServiceCategoryRef"
      serviceCandidate:
        type: "array"
        description: "List of service candidates associated with this category"
        items:
          $ref: "#/definitions/ServiceCandidateRef"
      validFor:
        description: "The period for which the category is valid"
        $ref: "#/definitions/TimePeriod"
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class"
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name"
    description: "The (service) category resource is used to group service candidates\
      \ in logical containers. Categories can contain other categories."
    example:
      isRoot: true
      lifecycleStatus: "lifecycleStatus"
      serviceCandidate:
      - '@referredType': "@referredType"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        '@schemaLocation': "http://example.com/aeiou"
        version: "version"
      - '@referredType': "@referredType"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        '@schemaLocation': "http://example.com/aeiou"
        version: "version"
      validFor:
        startDateTime: "2000-01-23T04:56:07.000+00:00"
        endDateTime: "2000-01-23T04:56:07.000+00:00"
      '@type': "@type"
      description: "description"
      version: "version"
      parentId: "parentId"
      '@baseType': "@baseType"
      lastUpdate: "2000-01-23T04:56:07.000+00:00"
      name: "name"
      id: "id"
      href: "href"
      category:
      - '@referredType': "@referredType"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        '@schemaLocation': "http://example.com/aeiou"
      - '@referredType': "@referredType"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        '@schemaLocation': "http://example.com/aeiou"
      '@schemaLocation': "http://example.com/aeiou"
  ServiceCategory_Create:
    type: "object"
    required:
    - "name"
    properties:
      description:
        type: "string"
        description: "Description of the category"
      isRoot:
        type: "boolean"
        description: "If true, this Boolean indicates that the category is a root\
          \ of categories"
      lastUpdate:
        type: "string"
        format: "date-time"
        description: "Date and time of the last update"
      lifecycleStatus:
        type: "string"
        description: "Used to indicate the current lifecycle status"
      name:
        type: "string"
        description: "Name of the category"
      parentId:
        type: "string"
        description: "Unique identifier of the parent category"
      version:
        type: "string"
        description: "ServiceCategory version"
      category:
        type: "array"
        description: "List of child categories in the tree for in this category"
        items:
          $ref: "#/definitions/ServiceCategoryRef"
      serviceCandidate:
        type: "array"
        description: "List of service candidates associated with this category"
        items:
          $ref: "#/definitions/ServiceCandidateRef"
      validFor:
        description: "The period for which the category is valid"
        $ref: "#/definitions/TimePeriod"
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class"
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name"
    description: "The (service) category resource is used to group service candidates\
      \ in logical containers. Categories can contain other categories.\nSkipped properties:\
      \ id,href"
    example:
      isRoot: true
      lifecycleStatus: "lifecycleStatus"
      serviceCandidate:
      - '@referredType': "@referredType"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        '@schemaLocation': "http://example.com/aeiou"
        version: "version"
      - '@referredType': "@referredType"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        '@schemaLocation': "http://example.com/aeiou"
        version: "version"
      validFor:
        startDateTime: "2000-01-23T04:56:07.000+00:00"
        endDateTime: "2000-01-23T04:56:07.000+00:00"
      '@type': "@type"
      description: "description"
      version: "version"
      parentId: "parentId"
      '@baseType': "@baseType"
      lastUpdate: "2000-01-23T04:56:07.000+00:00"
      name: "name"
      category:
      - '@referredType': "@referredType"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        '@schemaLocation': "http://example.com/aeiou"
      - '@referredType': "@referredType"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        '@schemaLocation': "http://example.com/aeiou"
      '@schemaLocation': "http://example.com/aeiou"
  ServiceCategory_Update:
    type: "object"
    properties:
      description:
        type: "string"
        description: "Description of the category"
      isRoot:
        type: "boolean"
        description: "If true, this Boolean indicates that the category is a root\
          \ of categories"
      lifecycleStatus:
        type: "string"
        description: "Used to indicate the current lifecycle status"
      name:
        type: "string"
        description: "Name of the category"
      parentId:
        type: "string"
        description: "Unique identifier of the parent category"
      version:
        type: "string"
        description: "ServiceCategory version"
      category:
        type: "array"
        description: "List of child categories in the tree for in this category"
        items:
          $ref: "#/definitions/ServiceCategoryRef"
      serviceCandidate:
        type: "array"
        description: "List of service candidates associated with this category"
        items:
          $ref: "#/definitions/ServiceCandidateRef"
      validFor:
        description: "The period for which the category is valid"
        $ref: "#/definitions/TimePeriod"
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class"
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name"
    description: "The (service) category resource is used to group service candidates\
      \ in logical containers. Categories can contain other categories.\nSkipped properties:\
      \ id,href,lastUpdate"
    example:
      isRoot: true
      lifecycleStatus: "lifecycleStatus"
      '@baseType': "@baseType"
      serviceCandidate:
      - '@referredType': "@referredType"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        '@schemaLocation': "http://example.com/aeiou"
        version: "version"
      - '@referredType': "@referredType"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        '@schemaLocation': "http://example.com/aeiou"
        version: "version"
      validFor:
        startDateTime: "2000-01-23T04:56:07.000+00:00"
        endDateTime: "2000-01-23T04:56:07.000+00:00"
      '@type': "@type"
      name: "name"
      description: "description"
      category:
      - '@referredType': "@referredType"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        '@schemaLocation': "http://example.com/aeiou"
      - '@referredType': "@referredType"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        '@schemaLocation': "http://example.com/aeiou"
      '@schemaLocation': "http://example.com/aeiou"
      version: "version"
      parentId: "parentId"
  ServiceCategoryRef:
    type: "object"
    required:
    - "href"
    - "id"
    properties:
      id:
        type: "string"
        description: "Unique identifier of category"
      href:
        type: "string"
        description: "Hypertext Reference of the category"
      name:
        type: "string"
        description: "Name of the category."
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class"
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name"
      '@referredType':
        type: "string"
        description: "The actual type of the target instance when needed for disambiguation."
    description: "The (service) category resource is used to group service candidates\
      \ in logical containers. Categories can contain other categories."
    example:
      '@referredType': "@referredType"
      '@baseType': "@baseType"
      '@type': "@type"
      name: "name"
      id: "id"
      href: "href"
      '@schemaLocation': "http://example.com/aeiou"
...
  TimePeriod:
    type: "object"
    properties:
      endDateTime:
        type: "string"
        format: "date-time"
        description: "End of the time period, using IETC-RFC-3339 format"
      startDateTime:
        type: "string"
        format: "date-time"
        description: "Start of the time period, using IETC-RFC-3339 format. If you\
          \ define a start, you must also define an end"
    description: "A period of time, either as a deadline (endDateTime only) a startDateTime\
      \ only, or both"
    example:
      startDateTime: "2000-01-23T04:56:07.000+00:00"
      endDateTime: "2000-01-23T04:56:07.000+00:00"
...
  Error:
    type: "object"
    required:
    - "code"
    - "reason"
    properties:
      code:
        type: "integer"
        description: "Application relevant detail, defined in the API or a common\
          \ list."
      reason:
        type: "string"
        description: "Explanation of the reason for the error which can be shown to\
          \ a client user."
      message:
        type: "string"
        description: "More details and corrective actions related to the error which\
          \ can be shown to a client user."
      status:
        type: "integer"
        description: "HTTP Error code extension"
      referenceError:
        type: "string"
        format: "uri"
        description: "URI of documentation describing the error."
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class."
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name."
    description: "Used when an API throws an Error, typically with a HTTP error response-code\
      \ (3xx, 4xx, 5xx)"